Rename DisplayBox to QueryBox
